requirements.txt
python-dotenv
requests
cachetools